# specs/api_contracts.yaml
openapi: "3.1.0"
info:
  title: ChatBridge v2 API
  version: "1.0.0"
  description: "K-12 AI chat platform with third-party app integration. Fastify backend with auto-generated OpenAPI (L-001)."
servers:
  - url: /api/v1
    description: API v1
paths:
...
  /apps/register:
    post:
      summary: Register a third-party app
      description: Developer registers app with tool definitions, UI manifest, permissions, compliance metadata.
      tags: [Apps]
      operationId: registerApp
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name, description, tool_definitions, ui_manifest, permissions, compliance_metadata]
              properties:
                name: { type: string }
                description: { type: string }
                tool_definitions: { type: object, description: "MCP-compatible tool schemas" }
                ui_manifest: { type: object, properties: { url: { type: string }, width: { type: integer }, height: { type: integer } } }
                permissions: { type: object }
                compliance_metadata: { type: object }
                interaction_model: { type: string, enum: [single_user, turn_based, concurrent] }
      responses:
        "201":
          description: App registered with pending_review status
          content:
            application/json:
              schema:
                type: object
                properties:
                  app_id: { type: string, format: uuid }
                  status: { type: string, enum: [pending_review] }
        "422": { description: "Validation errors" }
      x-requirement: FR-APP-001
  /apps/{appId}/tools/{toolName}/invoke:
    post:
      summary: Invoke an app tool
      description: AI-initiated tool invocation. Routes through backend, never direct iframe↔LLM.
      tags: [Apps]
      operationId: invokeAppTool
      parameters:
        - name: appId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: toolName
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                parameters: { type: object }
                conversation_id: { type: string, format: uuid }
      responses:
        "200": { description: "Tool result" }
        "404": { description: "App or tool not found" }
        "408": { description: "Tool execution timeout (>5s)" }
        "422": { description: "Invalid parameters" }
      x-requirement: FR-APP-002
...
  /apps/instances/{instanceId}/state:
    put:
      summary: Update app instance state
      description: App sends state update via CBP. Validated, safety-checked, persisted.
      tags: [Apps, CBP]
      operationId: updateAppState
      parameters:
        - name: instanceId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [state]
              properties:
                state: { type: object }
      responses:
        "200": { description: "State persisted" }
        "422": { description: "Safety check failed" }
      x-requirement: FR-STATE-001

    get:
      summary: Get app instance state
      description: Retrieve last known state for reconnection.
      tags: [Apps]
      operationId: getAppState
      parameters:
        - name: instanceId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: App state
          content:
            application/json:
              schema:
                type: object
                properties:
                  instance_id: { type: string }
                  status: { type: string }
                  state: { type: object }
      x-requirement: FR-STATE-003
...
  /admin/apps/{appId}/suspend:
    post:
      summary: Suspend app district-wide
      tags: [Admin]
      operationId: suspendApp
      responses:
        "200": { description: "App suspended, instances terminated, teachers notified" }
        "403": { description: "Requires district_admin role" }
      x-requirement: FR-CTRL-008
  /admin/safety-events:
    get:
      summary: Safety event audit trail
      tags: [Admin]
      operationId: getSafetyEvents
      parameters:
        - name: district_id
          in: query
          schema: { type: string, format: uuid }
        - name: start_date
          in: query
          schema: { type: string, format: date }
      responses:
        "200": { description: "Paginated safety events" }
      x-requirement: FR-OBS-003
  /admin/audit-trail:
    get:
      summary: FERPA audit trail
      tags: [Admin]
      operationId: getAuditTrail
      parameters:
        - name: student_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200": { description: "Paginated audit events" }
      x-requirement: FR-COMP-003
  /admin/tool-invocations:
    get:
      summary: Tool invocation log
      tags: [Admin]
      operationId: getToolInvocations
      parameters:
        - name: app_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200": { description: "Paginated tool invocation history" }
      x-requirement: FR-OBS-002

  # === Consent ===
  /consent/request:
    post:
      summary: Send parental consent request
      tags: [Consent]
      operationId: requestConsent
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [student_id, parent_email]
              properties:
                student_id: { type: string, format: uuid }
                parent_email: { type: string, format: email }
      responses:
        "200": { description: "Consent request sent" }
      x-requirement: FR-COMP-001
  /consent/delete-request:
    post:
      summary: Request student data deletion
      tags: [Consent]
      operationId: requestDeletion
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [student_id]
              properties:
                student_id: { type: string, format: uuid }
      responses:
        "200": { description: "Deletion request accepted, completes within 30 days" }
      x-requirement: FR-COMP-002

  # === Health ===
...